refactor(recurrence): split recurrence-handler into focused modules
Breaks the 708-line recurrence-handler.ts into five modules along its existing seams:
floating-time, series-helpers, generate-recurring-events, update-recurring-event,
delete-recurring-event. Acyclic deps (floating-time + series-helpers -> generate;
series-helpers -> update/delete). Repoints index/plugin/test imports at the concrete
modules (no re-export barrel); cross-module helpers gained export. Renames the handler
test to recurring-mutations.test.ts. Behavior-preserving (950 tests, build, type-check green).
