+++ docs/logs/2026-06-13.md
@@ -47,3 +47,19 @@
 - **[recurrence handler — delete dispatch]**: Converted `deleteRecurringEvent`'s `switch (scope)` to a `scopedDeleteHandlers` lookup map (`deleteThisScope`/`deleteFollowingScope`/`deleteAllScope`), mirroring `updateRecurringEvent`'s `scopedUpdateHandlers`. Both functions now dispatch the same scope union the same way and share the single `Invalid scope` throw shape. Behavior-preserving; pinned by the existing per-scope `events`/`updated`/`added`/`deleted` assertions.
 - **[recurrence handler — replaceStoredOverride]**: Changed from 5 positional args (three adjacent same-typed `CalendarEvent`) to a single `ReplaceStoredOverrideArgs` object param, matching the `ScopedUpdateContext` pattern. Removes the call-site transposition risk the types could not catch.
 - **[engine tests — dedupe]**: Added a `renderRecurrenceEngine(events)` helper in `use-calendar-engine.test.ts` that mounts the engine with the recurrence plugin and fresh callback spies; migrated all 11 recurring-event tests to it. Cut fallow's flagged duplication from 275 lines (15 clone groups) to 154 (9). Residual clones are pre-existing patterns in touched files (nav-range, range-query, provider blocks past my hunks) and explicit act/assert bodies, left as-is (out of scope / readability).
+
+## Changes (recurrence-handler split)
+
+- **[recurrence utils — module split]**: Broke the 708-line `recurrence-handler.ts` into five focused modules under `utils/` along its existing seams: `floating-time.ts` (RRule floating-time conversion), `series-helpers.ts` (shared series primitives: `isRecurringEvent`, `getEventParentUID`, `findBaseEventIndex`, `getOccurrenceStartISO`, `getSeriesTerminationDate`, `addExdateToBaseEvent`), `generate-recurring-events.ts`, `update-recurring-event.ts`, and `delete-recurring-event.ts`. Dependency DAG: floating-time + series-helpers ← generate; series-helpers ← update, delete. No cycles. Deleted the original file (no re-export barrel, per no-barrel rule).
+- **[importers]**: Repointed `index.ts`, `recurrence-plugin.tsx`, and the four util test files at the concrete new modules. Helpers crossing module boundaries gained `export` (were module-private `const`).
+- **[tests]**: `git mv recurrence-handler.test.ts → recurring-mutations.test.ts` and renamed its top-level describe. No test files created; imports repointed.
+
+## Files Modified (recurrence-handler split)
+
+- Created `packages/recurrence/src/utils/{floating-time,series-helpers,generate-recurring-events,update-recurring-event,delete-recurring-event}.ts`; deleted `recurrence-handler.ts`.
+- `packages/recurrence/src/index.ts`, `packages/recurrence/src/recurrence-plugin.tsx` - concrete import paths.
+- `packages/recurrence/src/utils/{recurring-mutations,crud-operations,calendar-integration,recurrence-timezone}.test.ts` - import paths (+ rename of the first).
+
+## Notes (recurrence-handler split)
+
+- Behavior-preserving move verified by the full suite (164 recurrence + 786 calendar pass), all 6 packages type-check, build green, fallow dead-code 0 / complexity 0 / duplication unchanged.
